Fixed Websocket multiple connections
Fixed a huge issue (with weird shenanigans, because the socketioclient library is missing an essential feature) that could have caused unintended behaviors when trying to connect multiple times to a down server. When the server would go back online, all the clients would try and connect to the server, leading to multiple clients connected to the server at the same time on one instance of the game/plugin.
This is fixed, now only one client can be connected in a normal use case.