Various changes
Started adding GlamourerIPC_Caller, for now it does nothing, this is just for future use.
Added debug logging for permissions received.
Rewrote PermissionsBuilder so that GlamourerPermissions and PenumbraPermissions can be null.
Changed the "SendPermissions" function to send a cleaned permissions list.
Added a timer for the "Automatically request their permissions".